Restructure usb_cam code into more digestible pieces
