+++ src/usb_cam_node.cpp
@@ -28,6 +28,7 @@
 
 
 #include "usb_cam/usb_cam_node.hpp"
+#include "usb_cam/utils.hpp"
 
 #include <sstream>
 // #include <std_srvs/srv/Empty.h>
@@ -125,22 +126,22 @@ void UsbCamNode::init()
     image_width_, image_height_, io_method_name_.c_str(),
     pixel_format_name_.c_str(), color_format_name_.c_str(), framerate_);
   // set the IO method
-  UsbCam::io_method io_method = UsbCam::io_method_from_string(io_method_name_);
-  if (io_method == UsbCam::IO_METHOD_UNKNOWN) {
+  io_method io_method = usb_cam::utils::io_method_from_string(io_method_name_);
+  if (io_method == usb_cam::utils::IO_METHOD_UNKNOWN) {
     RCLCPP_ERROR_ONCE(this->get_logger(), "Unknown IO method '%s'", io_method_name_.c_str());
     rclcpp::shutdown();
     return;
   }
   // set the pixel format
-  UsbCam::pixel_format pixel_format = UsbCam::pixel_format_from_string(pixel_format_name_);
-  if (pixel_format == UsbCam::PIXEL_FORMAT_UNKNOWN) {
+  pixel_format pixel_format = usb_cam::utils::pixel_format_from_string(pixel_format_name_);
+  if (pixel_format == usb_cam::utils::PIXEL_FORMAT_UNKNOWN) {
     RCLCPP_ERROR_ONCE(this->get_logger(), "Unknown pixel format '%s'", pixel_format_name_.c_str());
     rclcpp::shutdown();
     return;
   }
   // set the color format
-  UsbCam::color_format color_format = UsbCam::color_format_from_string(color_format_name_);
-  if (color_format == UsbCam::COLOR_FORMAT_UNKNOWN) {
+  color_format color_format = usb_cam::utils::color_format_from_string(color_format_name_);
+  if (color_format == usb_cam::utils::COLOR_FORMAT_UNKNOWN) {
     RCLCPP_ERROR_ONCE(this->get_logger(), "Unknown color format '%s'", color_format_name_.c_str());
     rclcpp::shutdown();
     return;
